Created separate thread for data analysis work
Running a simple k-means clustering algorithm example can be done by
doing any GET request to the /sequence Servlet.

WorkThread.java is a thread that gets started with the server, and looks
for JSON objects in a queue. Whenever it gathers an object from the
queue it creates an Instance object (from the JavaML library) and adds
it to a Dataset object (on which JavaML can run algorithms).